Move string permissions to old model
