+++ irohad/model/generators/impl/command_generator.cpp
@@ -17,7 +17,7 @@
 #include "model/commands/set_quorum.hpp"
 #include "model/commands/subtract_asset_quantity.hpp"
 #include "model/commands/transfer_asset.hpp"
-#include "validators/permissions.hpp"
+#include "model/permissions.hpp"
 
 using namespace generator;
 using namespace shared_model::permissions;
